global: fix hepcrawl build
* Re-factored: renamed `fix_venv_tmp_rights.c` to `fix_rights.c`.
* Re-factored: renamed `fix_venv_tmp_rights` to `fix_rights`.
* Adds: permission to `code` and `tmp` directory for docker images.
* Re-factored: `docker_entrypoint.sh` to support `code` folder.
* Re-factored: `docker-compose.test.yml` fixed build issue.
		It was wrong linking of `scrapyd` service with `scrapy-deploy` service.
