+++ docker-compose.test.yml
@@ -56,12 +56,12 @@ services:
     command: bash -c "rm -f twistd.pid && exec scrapyd"
     links:
       - celery
-      - scrapyd_deploy
 
   scrapyd_deploy:
     <<: *service_base
     command: bash -c "sleep 8 && scrapyd-deploy"  # make sure that the scrapyd is up
-    links: []
+    links:
+      - scrapyd
 
   ftp_server:
     image: stilliard/pure-ftpd:hardened
